requirements.txt
beautifulsoup4
duckduckgo_search
openai
python-dotenv
gradio